Refactor GestureOptions
