+++ Sources/MapboxMaps/MapView/Configuration/MapConfig.swift
@@ -2,8 +2,6 @@ import Foundation
 
 /// `MapConfig` is the structure used to configure the map with a set of capabilities
 public struct MapConfig: Equatable {
-    /// Used to configure the gestures on the map
-    public var gestures: GestureOptions = GestureOptions()
 
     /// Used to configure the camera of the map
     public var camera: MapCameraOptions = MapCameraOptions()
